[Storage][Blob]Move methods to corresponding file (#8949)
* [Storage][Blob]Move methods to corresponding file
